pyproject.toml
[tool.poetry]
name = "hassqitto"
version = "0.1.0"
description = "Home-Assistant MQTT Device"
authors = ["Harshad Sharma <harshad@sharma.io>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "hassquitto", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
APScheduler = "^3.10.1"
python-slugify = "^8.0.1"
asgiref = "^3.7.2"
asyncgnostic = "^0.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.12"
mkdocstrings = {extras = ["python"], version = "^0.21.2"}
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
tox = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
